+++ Worker.Objects/Resources/Crystal.cs
@@ -9,11 +9,11 @@ namespace Worker.Objects.Resources
 {
     public class Crystal : BasicResource
     {
-        public override double BasicProduction => 75;
+        public override int BasicProduction => 75;
 
         public Crystal(string amount) : base(amount) { }
 
-        public Crystal(double amount) : base(amount) { }
+        public Crystal(int amount) : base(amount) { }
 
     }
 }
