+++ Worker.Objects/Resources/Metal.cs
@@ -9,10 +9,10 @@ namespace Worker.Objects.Resources
 {
     public class Metal : BasicResource
     {
-        public override double BasicProduction => 150;
+        public override int BasicProduction => 150;
 
         public Metal(string amount) : base(amount) { }
 
-        public Metal(double amount) : base(amount) { }
+        public Metal(int amount) : base(amount) { }
     }
 }
